requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.2
cryptography==2.4.2
Django==2.1.7
django-filetransfers==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
djangorestframework-datatables==0.5.0
idna==2.8
netaddr==0.7.19
paramiko==2.4.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-magic==0.4.15
pytz==2018.9
scp==0.13.2
six==1.12.0
